Introduce a redhat_jira module
